Misc. improvements

+++ include/naive.h
@@ -1,1 +1,1 @@
-void naive(uint8_t[8][8], int[8][8]);
+void naive(int[8][8], int[8][8]);
